refactor(triangle): unify Triangle dataclass usage
- remove duplicate Triangle from engine module
- import models.Triangle across CLI, executor, and tests
- switch attribute access to leg_ab/leg_bc/leg_ac
